Analytics/requirements.txt
flask
flask-restful
flask_wtf
pymysql
psycopg2
flask-sqlalchemy
flask-migrate
flask-script
flask-cors
numpy
geoalchemy2
scipy
scikit-learn
fbprophet
holidays==0.9.8
geojson
shapely
bcrypt
flask-jwt-extended
pytest
logging
sendgrid
celery==v4.3.0rc1
apscheduler
